pyproject.toml
[tool.poetry]
name = "nft-data"
version = "0.1.0"
description = ""
authors = ["rarepepi <pepimartinez@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "^1.8.1"
requests = "^2.27.1"
pandas = "^1.4.2"
web3 = "^5.29.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
